Corrected behavior of `--overwrite` flag of `new-project` (#10)
The `new-project` command does not overwrite any existing files unless the `--overwrite` flag is used. However, the flag should not really be needed when the directory already exists, in case the directory is empty or doesn’t have the files we’ll generate.

Also, the command now better handles the error output of the underlying Strongback utility, looking specifically for an error message about an existing file and generates a more useful error message than what the utility generates. Any error message that doesn’t match these specific attributes is printed as it.

None of this changes the need to use the `--overwrite` flag, but when that flag is used the command behaves much better.